pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
exclude = ["Examples", "imgs"]  # exclude packages matching these glob patterns (empty by default)

[project]
name = "MAFE"
version = "0.0.1"
authors=[{name="Po-han Li", email="pohanli@utexas.edu"}]
description = """A package for learning basis functions over arbitrary function sets in multi-agent networks."""
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "torch>=2.0",
    "tqdm",
    "tensorboard",
    "numpy<=1.26.4"
]

[tool.poetry]
name = "MAFE"
version = "0.0.1"
authors = ["Po-han Li <pohanli@utexas.edu>"]
packages = [{include = "FunctionEncoder"}]
description = "Multi-agent function encoder"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
ruff = "*"
numpy = "<=1.26.4"
pandas = "*"
torch = ">=2.0"
tqdm = "*"
matplotlib = "*"
torchvision = "*"
tensorboard = "*"
hydra-core = "1.3.2"
hydra-joblib-launcher = "1.2.0"
omegaconf = "*"

[project.urls]
homepage = "https://github.com/UTAustin-SwarmLab/MultiAgentFunctionEncoder"
